RF: Moved ribbon wf (from bold to anat) and reformatted
