+++ nibabies/workflows/base.py
@@ -478,7 +478,7 @@ tasks and sessions), the following preprocessing was performed.
                     ('outputnode.subject_id', 'inputnode.subject_id'),
                     ('outputnode.t1w2fsnative_xfm', 'inputnode.t1w2fsnative_xfm'),
                     ('outputnode.fsnative2t1w_xfm', 'inputnode.fsnative2t1w_xfm'),
-                    ('outputnode.surfaces', 'inputnode.anat_giftis'),
+                    ('outputnode.anat_ribbon', 'inputnode.anat_ribbon'),
                 ]),
             ])
             # fmt: on
